requirements.txt
langchain
langchain-openai
openai
python-dotenv
requests
beautifulsoup4
gradio
